requirements.txt
Flask==3.1.1
flask_session==0.8.0
llama_index==0.12.52
llama-index-llms-openai>=0.1.0
llama-index-llms-google-genai>=0.1.0
openai==1.97.1
pandas>=2.0.0
numpy>=1.24.0
phoenix==0.9.1
protobuf==6.31.1
python-dotenv==1.1.1
requests>=2.31.0
google-generativeai>=0.3.0
arize-phoenix>=4.0.0
openinference-instrumentation-llama-index>=2.0.0